+++ src/cqfantasy/house/__init__.py
@@ -5,6 +5,8 @@ from .TileGenerator import TileGenerator
 from .House import House
 from .tudor_wall import tudor_wall
 from .WallTudor import WallTudor
+from .WallStuccoBrick import WallStuccoBrick
+from .WallTudorPaneling import WallTudorPaneling
 from .StuccoBrickBody import StuccoBrickBody
 from .TudorBody import TudorBody
 from .TudorSplitBody import TudorSplitBody